Update `agb`'s position and description (#61)
When it was first added, `agb` was definitely more of a library than a toolkit. But now I would consider it more of a toolkit / framework with significantly more built-in.

I've also update the URL to link to the website itself rather than the github page because I feel like we would want to direct people to the book rather than to the repo.